Test with button and CSS classes
