+++ src/routes/Home.svelte
@@ -1,9 +1,14 @@
 <script>
-	import MyButton from "$lib/components/MyButton.svelte";
+	import DplButton from "$lib/components/DplButton.svelte";
 
 </script>
 <div class="container mx-auto">
   <h1>Welcome DPL Glic</h1>
   <p class="mb-4">This is a simple Svelte application to track your weight and glucose levels.</p>
   <p class="mb-4">Use the navigation bar to access different sections of the app.</p>
+  <hr class="my-4" />
+
+  <DplButton>
+    Test Button
+  </DplButton>
 </div>
